Gif like button added.
GifItem - isLiked is a state. This commit shows that even though props are immutable, we can use states. Even though states get updated, we do not change props and litho supports this well.

I have added PreferenceLikeStore which stores whether the gif is liked or not. In GifItemView, we are passing GifCallback as a prop to update prefreences. This prop is optional so we can also pass a null value.

Signed-off-by: Jay Rambhia <jayrambhia777@gmail.com>