Full screen navigation added.
To show navigation, when the user clicks on an image, it triggers a callback and we create a FullScreenComponent and add it to our root view. And when the back button is pressed, we add HomeComponent to root. So we not necessarily require Activity navigation.

In GifImageView, we are passing an optional prop isFullScreen. If it's true, we pass the ratio of width/height to measureWithAspectRatio else we pass 1 to show square images. Also, this prop is used in onMount where we decide if we want to use original image or smaller image to load via glide.

Signed-off-by: Jay Rambhia <jayrambhia777@gmail.com>