state synchronization added.
Updated FullScreenComponent key to gifId + time. If the key is gif id, for the same gif we get same component and its @OnCreateInitialState is not called. So if we have already opened a gif before and we update like state and open the gif again the state is not updated.

Signed-off-by: Jay Rambhia <jayrambhia777@gmail.com>